refactor some ucan stuff
