+++ src/ucan/index.ts
@@ -1,0 +1,3 @@
+export * from './types'
+export * from './token'
+export * as dictionary from './dictionary'
